# api/openapi.yaml
openapi: 3.1.0
info:
  title: Nepal Heritage API
  version: 0.1.0
  description: >
    Accounts and sessions for the Nepal Heritage Flutter app. This contract is
    the source of truth: the server and the Dart client both evolve from it.
servers:
  - url: http://127.0.0.1:3000
    description: Local development
tags:
  - name: health
  - name: auth
  - name: account
paths:
  /health:
    get:
      tags: [health]
      operationId: getHealth
      summary: Liveness probe
      responses:
        "200":
          description: Service is up
          content:
            application/json:
              schema:
                type: object
                required: [status, service]
                properties:
                  status:
                    type: string
                    const: ok
                  service:
                    type: string
  /api/v1/auth/register:
    post:
      tags: [auth]
      operationId: register
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created; the body carries a session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Sign in with an email or a phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Signed in; the body carries a session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/reset-password:
    post:
      tags: [auth]
      operationId: requestPasswordReset
      summary: Request a password reset link
      description: >
        Always answers 202 with the same message — whether an account exists
        must not leak through this endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "202":
          description: Request accepted (regardless of whether the identifier exists)
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/me:
    get:
      tags: [account]
      operationId: getMe
      summary: Current session's user (session restore)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The signed-in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, invalid or expired credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The email is already registered
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Stable machine-readable code
        message:
          type: string
          description: Human-readable text safe to show in the UI
    AuthUser:
      type: object
      required: [id, fullName, email, completedOnboarding]
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
          format: email
        completedOnboarding:
          type: boolean
          description: False sends the client through the intro carousel once
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        token:
          type: string
          description: Bearer JWT, valid for 30 days
    UserResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
    RegisterRequest:
      type: object
      required: [fullName, gender, dialCode, phone, email, password]
      properties:
        fullName:
          type: string
        address:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        nationalId:
          type: string
        passportNumber:
          type: string
        dialCode:
          type: string
          example: "+977"
        phone:
          type: string
          example: "9801234567"
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        country:
          type: string
          minLength: 2
          maxLength: 2
          default: NP
          description: ISO 3166-1 alpha-2
        audioGuideLanguage:
          type: string
          default: en
          example: ne
    LoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: Email address or phone number (local, or with country prefix)
        password:
          type: string
          minLength: 6
        rememberMe:
          type: boolean
          default: false
    ResetPasswordRequest:
      type: object
      required: [identifier]
      properties:
        identifier:
          type: string
          description: Email address or phone number